+++ test/debug_plugin_method_channel_test.dart
@@ -1,27 +1,27 @@
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:debug_plugin/debug_plugin_method_channel.dart';
+// import 'package:flutter/services.dart';
+// import 'package:flutter_test/flutter_test.dart';
+// import 'package:debug_plugin/debug_plugin_method_channel.dart';
 
-void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+// void main() {
+//   TestWidgetsFlutterBinding.ensureInitialized();
 
-  MethodChannelDebugPlugin platform = MethodChannelDebugPlugin();
-  const MethodChannel channel = MethodChannel('debug_plugin');
+//   MethodChannelDebugPlugin platform = MethodChannelDebugPlugin();
+//   const MethodChannel channel = MethodChannel('debug_plugin');
 
-  setUp(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
-      channel,
-      (MethodCall methodCall) async {
-        return '42';
-      },
-    );
-  });
+//   setUp(() {
+//     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
+//       channel,
+//       (MethodCall methodCall) async {
+//         return '42';
+//       },
+//     );
+//   });
 
-  tearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
-  });
+//   tearDown(() {
+//     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
+//   });
 
-  test('getPlatformVersion', () async {
-    expect(await platform.getPlatformVersion(), '42');
-  });
-}
+//   test('getPlatformVersion', () async {
+//     expect(await platform.getPlatformVersion(), '42');
+//   });
+// }